Binary Search Tree Completion
Finished the Binary Search Tree tutorial I found on YouTube. I would hate to have to constantly write that from scratch, it took a lot of time.
The tutorial was done recursively so starting at the next sesssion, I'm going to challenge myself to recreate the tutorial iteratively and try not to break the damn thing.
Also I think I'm going to make it a point to save the address of the tutorial I'm using just for reference.